feat: Day 01 - Part a

+++ 2024/golang/src/template/main.go
@@ -1,9 +1,9 @@
 package main
 
-func day01a(filename string) float64 {
+func day01a(filename string) int {
 	return 0
 }
 
-func day01b(filename string) float64 {
+func day01b(filename string) int {
 	return 0
 }
